+++ tests/test_base.py
@@ -60,23 +60,23 @@ def test_query_params():
     assert r['content'][0:10] == b'RIFF$\x07\x06\x00WA'
 
 def test_event_type_data():
-    misty_robot.subscribe_event_type("BatteryCharge", event_name="Battery Event")
+    misty_robot.subscribe_event("BatteryCharge", event_name="Battery Event")
     time.sleep(1)
     data = misty_robot.get_event_data("Battery Event")
-    misty_robot.unsubscribe_event_type("Battery Event")
+    misty_robot.unsubscribe_event("Battery Event")
     assert data['status'] == 'Success'
 
 def test_event_type_log():
-    misty_robot.subscribe_event_type("BatteryCharge", event_name="Battery Event")
+    misty_robot.subscribe_event("BatteryCharge", event_name="Battery Event")
     time.sleep(1)
     data = misty_robot.get_event_log("Battery Event")
-    misty_robot.unsubscribe_event_type("Battery Event")
+    misty_robot.unsubscribe_event("Battery Event")
     assert data['status'] == 'Success'
 
 def test_event_type_unsubscription():
-    misty_robot.subscribe_event_type("BatteryCharge", event_name="Battery Event")
+    misty_robot.subscribe_event("BatteryCharge", event_name="Battery Event")
     time.sleep(1)
     data = misty_robot.get_event_data("Battery Event")
-    misty_robot.unsubscribe_event_type("Battery Event")
+    misty_robot.unsubscribe_event("Battery Event")
     nodata = misty_robot.get_event_data("Battery Event")
     assert data['status'] == 'Success' and nodata['status'] == 'Failed'